+++ trace.go
@@ -1,0 +1,55 @@
+package connect
+
+import (
+	"context"
+
+	log "github.com/sirupsen/logrus"
+
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
+	"go.opentelemetry.io/otel/propagation"
+	"go.opentelemetry.io/otel/sdk/resource"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+)
+
+// InitTraceProvider configures an OpenTelemetry exporter and trace provider
+func InitTraceProvider(token, collectorURL, serviceName string, traceRatio float64) *sdktrace.TracerProvider {
+	headers := map[string]string{
+		"signoz-access-token": token,
+	}
+
+	exporter, err := otlptrace.New(
+		context.Background(),
+		otlptracegrpc.NewClient(
+			otlptracegrpc.WithInsecure(),
+			otlptracegrpc.WithEndpoint(collectorURL),
+			otlptracegrpc.WithHeaders(headers),
+		),
+	)
+
+	if err != nil {
+		log.Fatal(err)
+	}
+	resources, err := resource.New(
+		context.Background(),
+		resource.WithAttributes(
+			attribute.String("service.name", serviceName),
+		),
+	)
+	if err != nil {
+		log.Printf("Could not set resources: %v", err)
+	}
+
+	traceProvider := sdktrace.NewTracerProvider(
+		sdktrace.WithSampler(sdktrace.TraceIDRatioBased(traceRatio)),
+		sdktrace.WithBatcher(exporter),
+		sdktrace.WithResource(resources),
+	)
+
+	otel.SetTracerProvider(traceProvider)
+	otel.SetTextMapPropagator(propagation.NewCompositeTextMapPropagator(propagation.TraceContext{}, propagation.Baggage{}))
+
+	return traceProvider
+}